Major reorganization to enable FireSim-as-a-library
