+++ sim/firesim-lib/src/main/scala/configs/SimConfigs.scala
@@ -1,12 +1,11 @@
-package firesim.firesim
+//See LICENSE for license details.
+package firesim.configs
 
 import freechips.rocketchip.config.{Parameters, Config, Field}
 
 import midas.{EndpointKey}
 import midas.widgets.{EndpointMap}
 import midas.models._
-
-import testchipip.{WithBlockDevice}
 
 import firesim.endpoints._
 
@@ -143,7 +142,7 @@ class LBP32R32W3Div extends Config(
   new WithDefaultMemModel(3))
 
 // DDR3 - FCFS models.
-class FCFS16GBQuadRank extends Config(new WithDDR3FIFOMAS(8) ++ new FireSimConfig)
+class FCFS16GBQuadRank extends Config(new WithDDR3FIFOMAS(8) ++ new WithDefaultMemModel)
 class FCFS16GBQuadRankLLC4MB extends Config(
   new WithLLCModel(4096, 8) ++
   new FCFS16GBQuadRank)
@@ -164,45 +163,45 @@ class FRFCFS16GBQuadRankLLC4MB3Div extends Config(
   new FRFCFS16GBQuadRank(3)
 )
 
-/*******************************************************************************
-* Full PLATFORM_CONFIG Configurations. These set simulator parameters.
-*
-* In general, if you're adding or removing features from any of these, you
-* should CREATE A NEW ONE, WITH A NEW NAME. This is because the manager
-* will store this name as part of the tags for the AGFI, so that later you can
-* reconstruct what is in a particular AGFI. These tags are also used to
-* determine which driver to build.
-*******************************************************************************/
-class FireSimConfig extends Config(
-  new WithSerialWidget ++
-  new WithUARTWidget ++
-  new WithSimpleNICWidget ++
-  new WithBlockDevWidget ++
-  new WithDefaultMemModel ++
-  new WithTracerVWidget ++
-  new BasePlatformConfig)
-
-class FireSimClockDivConfig extends Config(
-  new WithDefaultMemModel(clockDivision = 2) ++
-  new FireSimConfig)
-
-class FireSimDDR3Config extends Config(
-  new FCFS16GBQuadRank ++
-  new FireSimConfig)
-
-class FireSimDDR3LLC4MBConfig extends Config(
-  new FCFS16GBQuadRankLLC4MB ++
-  new FireSimConfig)
-
-class FireSimDDR3FRFCFSConfig extends Config(
-  new FRFCFS16GBQuadRank ++
-  new FireSimConfig)
-
-class FireSimDDR3FRFCFSLLC4MBConfig extends Config(
-  new FRFCFS16GBQuadRankLLC4MB ++
-  new FireSimConfig)
-
-class FireSimDDR3FRFCFSLLC4MB3ClockDivConfig extends Config(
-  new FRFCFS16GBQuadRankLLC4MB3Div ++
-  new FireSimConfig)
-
+///*******************************************************************************
+//* Full PLATFORM_CONFIG Configurations. These set simulator parameters.
+//*
+//* In general, if you're adding or removing features from any of these, you
+//* should CREATE A NEW ONE, WITH A NEW NAME. This is because the manager
+//* will store this name as part of the tags for the AGFI, so that later you can
+//* reconstruct what is in a particular AGFI. These tags are also used to
+//* determine which driver to build.
+//*******************************************************************************/
+//class FireSimConfig extends Config(
+//  new WithSerialWidget ++
+//  new WithUARTWidget ++
+//  new WithSimpleNICWidget ++
+//  new WithBlockDevWidget ++
+//  new WithDefaultMemModel ++
+//  new WithTracerVWidget ++
+//  new BasePlatformConfig)
+//
+//class FireSimClockDivConfig extends Config(
+//  new WithDefaultMemModel(clockDivision = 2) ++
+//  new FireSimConfig)
+//
+//class FireSimDDR3Config extends Config(
+//  new FCFS16GBQuadRank ++
+//  new FireSimConfig)
+//
+//class FireSimDDR3LLC4MBConfig extends Config(
+//  new FCFS16GBQuadRankLLC4MB ++
+//  new FireSimConfig)
+//
+//class FireSimDDR3FRFCFSConfig extends Config(
+//  new FRFCFS16GBQuadRank ++
+//  new FireSimConfig)
+//
+//class FireSimDDR3FRFCFSLLC4MBConfig extends Config(
+//  new FRFCFS16GBQuadRankLLC4MB ++
+//  new FireSimConfig)
+//
+//class FireSimDDR3FRFCFSLLC4MB3ClockDivConfig extends Config(
+//  new FRFCFS16GBQuadRankLLC4MB3Div ++
+//  new FireSimConfig)
+//
