+++ sim/firesim-lib/src/main/scala/endpoints/SimpleNICWidget.scala
