+++ sim/firesim-lib/src/main/scala/endpoints/TracerVWidget.scala
