+++ sim/src/main/makefrag/midasexamples/Makefrag
@@ -19,8 +19,10 @@ OUTPUT_DIR    := $(base_dir)/output/$(PLATFORM)/$(name_tuple)
 VERILOG := $(GENERATED_DIR)/FPGATop.v
 HEADER  := $(GENERATED_DIR)/$(DESIGN)-const.h
 
-submodules = . midas \
-	$(addprefix target-rtl/firechip/rocket-chip/, . hardfloat chisel3 chisel3/chiselFrontend)
+submodules = . midas midas/targetutils firesim-lib \
+    $(addprefix target-rtl/firechip/, \
+		$(addprefix generators/, rocket-chip rocket-chip/hardfloat rocket-chip/macros) \
+		$(addprefix tools/, firrtl chisel3 chisel3/chiselFrontend))
 
 src_path = src/main/scala
 chisel_srcs = $(foreach submodule,$(submodules),$(shell find $(base_dir)/$(submodule)/$(src_path) -name "*.scala"))
